refactor: KakaoUserInfo 폴더 dto로 변경
