+++ agoda-clone/src/main/java/com/efub/agodaclone/user/controller/AuthController.java
@@ -1,6 +1,6 @@
 package com.efub.agodaclone.user.controller;
 
-import com.efub.agodaclone.user.domain.KakaoUserInfo;
+import com.efub.agodaclone.user.dto.KakaoUserInfo;
 import com.efub.agodaclone.user.domain.User;
 import com.efub.agodaclone.user.dto.AuthResponse;
 import com.efub.agodaclone.user.jwt.JwtProvider;
@@ -9,7 +9,6 @@ import com.efub.agodaclone.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
